feat: delegated composables to its correct folder; created translations to es, us, pt;
