+++ app/src/main/res/values/strings.xml
@@ -19,6 +19,9 @@
     <string name="no_tags_informed">No tags informed.</string>
     <string name="agent_details">Agent Details</string>
     <string name="weapon_skins">Skins</string>
+    <string name="dropdown_portuguese">Portuguese</string>
+    <string name="dropdown_english">English</string>
+    <string name="dropdown_spanish">Spanish</string>
 
     <!-- adicionadas para WeaponSingleScreen (agora em inglês) -->
     <string name="category">Category: %1$s</string>
